src/pyproject.toml
[tool.poetry]
name = "data_engineer_example_kafka"
version = "0.1.0"
description = ""
authors = ["mauza <mauza@pluralsight.com>"]

[tool.poetry.dependencies]
python = "^3.8"
confluent_kafka = "^1.5.0"
python-dotenv = "^0.14.0"
selenium = "^3.141.0"
beautifulsoup4 = "^4.9.3"
requests = "^2.24.0"
fastavro = "^1.0.0"
jsonschema = "^3.2.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
